requirements.txt
opencv-python
numpy
pillow
pytesseract
easyocr
pdfplumber
pypdf
sympy
scikit-image